+++ remember-me-controls.php
@@ -134,7 +134,7 @@ final class c2c_RememberMeControls extends c2c_RememberMeControls_Plugin_050 {
 				'default'  => '',
 				'datatype' => 'int',
 				'label'    => __( 'Remember Me duration', 'remember-me-controls' ),
-				'help'     => __( 'The number of <strong>hours</strong> a login with "Remember Me" checked will last. If not provided, then the WordPress default of 336 (i.e. two weeks) will be used. Do not include any commas. This value is ignored if "Remember forever?" is checked above.', 'remember-me-controls' )
+				'help'     => __( 'The number of <strong>hours</strong> a login with "Remember Me" checked will last. If not provided, then the WordPress default of 336 (i.e. two weeks) will be used. This value is ignored if "Remember forever?" is checked above.', 'remember-me-controls' )
 					. sprintf( '<br><em>%s</em>', __( 'NOTE: A change of this value only takes effect on subsequent logins.', 'remember-me-controls' ) ),
 			),
 			'disable_remember_me' => array(
@@ -156,6 +156,7 @@ final class c2c_RememberMeControls extends c2c_RememberMeControls_Plugin_050 {
 		add_action( 'login_head',                             array( $this, 'add_css' ) );
 		add_filter( 'login_footer',                           array( $this, 'add_js' ) );
 		add_action( $this->get_hook( 'post_display_option' ), array( $this, 'maybe_add_hr' ) );
+		add_filter( $this->get_hook( 'before_update_option' ), array( $this, 'sanitize_remember_me_duration' ) );
 		add_filter( 'login_form_defaults',                    array( $this, 'login_form_defaults' ) );
 
 		// Compat for BuddyPress Login Widget.
@@ -279,6 +280,24 @@ JS;
 		if ( 'remember_me_duration' === $opt ) {
 			echo "</tr><tr><td colspan='2'><div class='hr'>&nbsp;</div></td>\n";
 		}
+	}
+
+	/**
+	 * Sanitizes the value for the remember_me_duration setting.
+	 *
+	 * Basically it removes any commas.
+	 *
+	 * @since 1.9
+	 *
+	 * @param  array $options The options array prior to saving.
+	 * @return array
+	 */
+	public function sanitize_remember_me_duration( $options ) {
+		if ( ! empty( $options['remember_me_duration'] ) ) {
+			$options['remember_me_duration'] = str_replace( ',', '', $options['remember_me_duration'] );
+		}
+
+		return $options;
 	}
 
 	/**